+++ Cargo.toml
@@ -95,7 +95,6 @@ members = [
     "crates/primitives/",
     "crates/prune/db",
     "crates/prune/prune",
-    "crates/prune/static-files",
     "crates/prune/types",
     "crates/ress/protocol",
     "crates/ress/provider",
@@ -431,7 +430,6 @@ reth-primitives = { path = "crates/primitives", default-features = false }
 reth-primitives-traits = { path = "crates/primitives-traits", default-features = false }
 reth-provider = { path = "crates/storage/provider" }
 reth-prune = { path = "crates/prune/prune" }
-reth-prune-static-files = { path = "crates/prune/static-files" }
 reth-prune-types = { path = "crates/prune/types", default-features = false }
 reth-revm = { path = "crates/revm", default-features = false }
 reth-rpc = { path = "crates/rpc/rpc" }
