+++ crates/prune/prune/Cargo.toml
@@ -19,7 +19,6 @@ reth-errors.workspace = true
 reth-provider.workspace = true
 reth-tokio-util.workspace = true
 reth-config.workspace = true
-reth-prune-static-files.workspace = true
 reth-prune-types.workspace = true
 reth-primitives-traits.workspace = true
 reth-static-file-types.workspace = true
