+++ crates/prune/prune/src/segments/user/mod.rs
@@ -8,7 +8,7 @@ mod storage_history;
 mod transaction_lookup;
 
 pub use account_history::AccountHistory;
-pub use bodies::BodiesAdapter;
+pub use bodies::Bodies;
 pub use merkle_change_sets::MerkleChangeSets;
 pub use receipts::Receipts;
 pub use sender_recovery::SenderRecovery;
