--- crates/prune/static-files/src/lib.rs
@@ -1,14 +1,0 @@
-//! Pruning for static files.
-//!
-//! This crate provides pruning functionality for data stored in static files.
-
-#![doc(
-    html_logo_url = "https://raw.githubusercontent.com/paradigmxyz/reth/main/assets/reth-docs.png",
-    html_favicon_url = "https://avatars0.githubusercontent.com/u/97369466?s=256",
-    issue_tracker_base_url = "https://github.com/paradigmxyz/reth/issues/"
-)]
-#![cfg_attr(docsrs, feature(doc_cfg))]
-#![cfg_attr(not(test), warn(unused_crate_dependencies))]
-
-mod bodies;
-pub use bodies::Bodies;
